Share config between workspace root and apps/prairielearn vitest
